Рефакторинг кода в create-big-photo, добавил функции открытия и закрытия большого фото
